Add DTO's to extract necessary info
